refactor: add osu-native calculator, abstract performance logic and upgrade attributes to double precision

+++ src/main/java/pe/nanamochi/banchus/entities/db/Stat.java
@@ -23,7 +23,7 @@ public class Stat implements Cloneable {
   private int performancePoints;
   private int playCount;
   private int playTime;
-  private float accuracy;
+  private double accuracy;
   private int highestCombo;
   private int totalHits;
   private int replayViews;
